[Vulkan] Fixed validation error related to an immediate command list semaphore which was signaled but not waited for

+++ runtime/RHI/D3D12/D3D12_Device.cpp
@@ -149,7 +149,7 @@ namespace Spartan
         SP_LOG_INFO("DirectX %s", level.c_str());
     }
 
-    void RHI_Device::Tick(const uint32_t frame_count)
+    void RHI_Device::Tick(const uint64_t frame_count)
     {
 
     }
